# src/Api/openapi/portfolio-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Portfolio API",
    "description": "Slice 01: manual balance entry to a net-worth chart. Money is a decimal string at scale 4.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/system/health": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "GetHealth",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/net-worth": {
      "get": {
        "tags": [
          "Net worth"
        ],
        "operationId": "GetNetWorth",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "operationId": "ListAccounts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "CreateAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{id}/balances": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "operationId": "ListBalances",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BalanceResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "RecordBalance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountResponse": {
        "required": [
          "id",
          "institutionName",
          "displayName",
          "currency",
          "openedOn",
          "closedOn",
          "latestBalance",
          "latestAsOfDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "institutionName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "openedOn": {
            "type": "string",
            "format": "date"
          },
          "closedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "latestBalance": {
            "type": [
              "null",
              "string"
            ],
            "description": "The newest live balance, as a decimal string; null if none recorded."
          },
          "latestAsOfDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
      "BalanceResponse": {
        "required": [
          "id",
          "accountId",
          "asOfDate",
          "balance",
          "currency",
          "sourceSystem",
          "sourceStrength",
          "observedAt",
          "supersedes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "asOfDate": {
            "type": "string",
            "format": "date"
          },
          "balance": {
            "type": "string",
            "description": "Decimal string at scale 4."
          },
          "currency": {
            "type": "string"
          },
          "sourceSystem": {
            "type": "string"
          },
          "sourceStrength": {
            "$ref": "#/components/schemas/SourceStrength"
          },
          "observedAt": {
            "type": "string",
            "format": "date-time"
          },
          "supersedes": {
            "type": [
              "null",
              "string"
            ],
            "description": "The balance this one restated, if any.",
            "format": "uuid"
          }
        }
      },
      "CreateAccountRequest": {
        "required": [
          "institutionName",
          "displayName",
          "openedOn"
        ],
        "type": "object",
        "properties": {
          "institutionName": {
            "type": "string",
            "description": "Institution name. Created if no institution has it yet."
          },
          "displayName": {
            "type": "string"
          },
          "openedOn": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "HealthResponse": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "NetWorthPoint": {
        "required": [
          "asOfDate",
          "netWorth",
          "status",
          "accountsInWindow",
          "accountsVerified",
          "accountsUnverified",
          "accountsCarried",
          "maxStalenessDays"
        ],
        "type": "object",
        "properties": {
          "asOfDate": {
            "type": "string",
            "format": "date"
          },
          "netWorth": {
            "type": [
              "null",
              "string"
            ],
            "description": "Decimal string; null when no account is verified - never read it as zero."
          },
          "status": {
            "$ref": "#/components/schemas/NetWorthStatus"
          },
          "accountsInWindow": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "accountsVerified": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "accountsUnverified": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "accountsCarried": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "maxStalenessDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "NetWorthResponse": {
        "required": [
          "from",
          "to",
          "stalenessDays",
          "points"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "stalenessDays": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Days after which a carried balance goes unverified (D-034).",
            "format": "int32"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NetWorthPoint"
            }
          }
        }
      },
      "NetWorthStatus": {
        "enum": [
          "verified",
          "partial",
          "unverified"
        ]
      },
      "RecordBalanceRequest": {
        "required": [
          "asOfDate",
          "balance"
        ],
        "type": "object",
        "properties": {
          "asOfDate": {
            "type": "string",
            "format": "date"
          },
          "balance": {
            "type": "string",
            "description": "Decimal string, at most 4 decimal places, e.g. \"1234.5678\"."
          }
        }
      },
      "SourceStrength": {
        "enum": [
          "statement",
          "export",
          "api",
          "scrape",
          "manual"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Api"
    },
    {
      "name": "Net worth"
    },
    {
      "name": "Accounts"
    }
  ]
}